--- src/vs/platform/userDataSync/common/keybindingsSyncIpc.ts
@@ -1,45 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { IChannel, IServerChannel } from 'vs/base/parts/ipc/common/ipc';
-import { Event } from 'vs/base/common/event';
-import { IUserDataSyncUtilService } from 'vs/platform/userDataSync/common/userDataSync';
-import { IStringDictionary } from 'vs/base/common/collections';
-import { URI } from 'vs/base/common/uri';
-import { FormattingOptions } from 'vs/base/common/jsonFormatter';
-
-export class UserDataSycnUtilServiceChannel implements IServerChannel {
-
-	constructor(private readonly service: IUserDataSyncUtilService) { }
-
-	listen(_: unknown, event: string): Event<any> {
-		throw new Error(`Event not found: ${event}`);
-	}
-
-	call(context: any, command: string, args?: any): Promise<any> {
-		switch (command) {
-			case 'resolveUserKeybindings': return this.service.resolveUserBindings(args[0]);
-			case 'resolveFormattingOptions': return this.service.resolveFormattingOptions(URI.revive(args[0]));
-		}
-		throw new Error('Invalid call');
-	}
-}
-
-export class UserDataSyncUtilServiceClient implements IUserDataSyncUtilService {
-
-	_serviceBrand: undefined;
-
-	constructor(private readonly channel: IChannel) {
-	}
-
-	async resolveUserBindings(userbindings: string[]): Promise<IStringDictionary<string>> {
-		return this.channel.call('resolveUserKeybindings', [userbindings]);
-	}
-
-	async resolveFormattingOptions(file: URI): Promise<FormattingOptions> {
-		return this.channel.call('resolveFormattingOptions', [file]);
-	}
-
-}
