+++ src/vs/platform/userDataSync/common/userDataSyncIpc.ts
@@ -3,9 +3,12 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IServerChannel } from 'vs/base/parts/ipc/common/ipc';
+import { IServerChannel, IChannel } from 'vs/base/parts/ipc/common/ipc';
 import { Event } from 'vs/base/common/event';
-import { IUserDataSyncService } from 'vs/platform/userDataSync/common/userDataSync';
+import { IUserDataSyncService, IUserDataSyncUtilService } from 'vs/platform/userDataSync/common/userDataSync';
+import { URI } from 'vs/base/common/uri';
+import { IStringDictionary } from 'vs/base/common/collections';
+import { FormattingOptions } from 'vs/base/common/jsonFormatter';
 
 export class UserDataSyncChannel implements IServerChannel {
 
@@ -30,3 +33,38 @@ export class UserDataSyncChannel implements IServerChannel {
 		throw new Error('Invalid call');
 	}
 }
+
+export class UserDataSycnUtilServiceChannel implements IServerChannel {
+
+	constructor(private readonly service: IUserDataSyncUtilService) { }
+
+	listen(_: unknown, event: string): Event<any> {
+		throw new Error(`Event not found: ${event}`);
+	}
+
+	call(context: any, command: string, args?: any): Promise<any> {
+		switch (command) {
+			case 'resolveUserKeybindings': return this.service.resolveUserBindings(args[0]);
+			case 'resolveFormattingOptions': return this.service.resolveFormattingOptions(URI.revive(args[0]));
+		}
+		throw new Error('Invalid call');
+	}
+}
+
+export class UserDataSyncUtilServiceClient implements IUserDataSyncUtilService {
+
+	_serviceBrand: undefined;
+
+	constructor(private readonly channel: IChannel) {
+	}
+
+	async resolveUserBindings(userbindings: string[]): Promise<IStringDictionary<string>> {
+		return this.channel.call('resolveUserKeybindings', [userbindings]);
+	}
+
+	async resolveFormattingOptions(file: URI): Promise<FormattingOptions> {
+		return this.channel.call('resolveFormattingOptions', [file]);
+	}
+
+}
+
